requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.14.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
black=21.4b2=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
bottleneck=1.3.2=py39hdd57654_1
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.2=pypi_0
certifi=2021.10.8=py39h06a4308_0
cffi=1.15.0=pypi_0
charset-normalizer=2.0.6=pypi_0
click=8.0.1=pypi_0
cloudpickle=2.0.0=pypi_0
cycler=0.10.0=pypi_0
cython=0.29.24=pypi_0
debugpy=1.5.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
detectron2=0.5+cu111=pypi_0
entrypoints=0.3=pypi_0
future=0.18.2=pypi_0
fvcore=0.1.5.post20210924=pypi_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.40.0=pypi_0
hydra-core=1.1.1=pypi_0
idna=3.2=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.8=pypi_0
ipykernel=6.6.1=pypi_0
ipython=7.31.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jsonschema=4.3.3=pypi_0
jupyter-client=7.1.0=pypi_0
jupyter-core=4.9.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.3.2=pypi_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libstdcxx-ng=9.3.0=hd4cf53a_17
markdown=3.3.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.9=pypi_0
nbconvert=6.4.0=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.4=pypi_0
notebook=6.4.6=pypi_0
numexpr=2.8.1=py39h6abb31d_0
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
nvidia-ml-py=11.450.51=pypi_0
nvitop=0.4.3=pypi_0
oauthlib=3.1.1=pypi_0
omegaconf=2.1.1=pypi_0
opencv-python=4.5.3.56=pypi_0
openssl=1.1.1l=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py39h8c16a72_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.9.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.2=pypi_0
pip=21.2.4=py37h06a4308_0
portalocker=2.3.2=pypi_0
prometheus-client=0.12.0=pypi_0
prompt-toolkit=3.0.24=pypi_0
protobuf=3.18.0=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pycparser=2.21=pypi_0
pydot=1.4.2=pypi_0
pygments=2.11.2=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.18.0=pypi_0
python=3.9.0=hdb3f193_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=pypi_0
readline=8.1=h27cfd23_0
regex=2021.8.28=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-learn=1.0.1=py39h51133e4_0
scipy=1.7.3=py39hc147768_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tabulate=0.8.9=pypi_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.12.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pypi_0
torch=1.9.0+cu111=pypi_0
torchvision=0.10.0+cu111=pypi_0
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pypi_0
typing-extensions=3.10.0.2=pypi_0
tzdata=2021a=h5d7bf9c_0
urllib3=1.26.7=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.2=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yolox=0.1.0=dev_0
zlib=1.2.11=h7b6447c_3
